Casi último cambios, category añadido

+++ src/controllers/offer.controller.ts
@@ -6,7 +6,8 @@ export class OfferController{
 
     static async getAll(req:Request, res:Response, next:NextFunction){
         try{
-            const offers = await OfferService.getAll()
+            const {title} = req.params
+            const offers = await OfferService.getAll(title as string)
             res.status(200).json(offers)
         }catch(error){
             next(error)
@@ -27,8 +28,10 @@ export class OfferController{
 
     static async save(req:Request, res:Response, next:NextFunction){
         try{
-            const userId = req.user.id
             const offerData = req.body
+            const userId = req.user?.id
+            if(!userId) throw new HttpException(400,'User creator ID is required') 
+            
             const newOffer = await OfferService.save(userId,offerData)
             res.status(201).json({message: 'Offer created successfully', newOffer})
         }catch(error){
@@ -39,8 +42,10 @@ export class OfferController{
     static async delete(req:Request, res:Response, next:NextFunction){
         try{
             const id = Number.parseInt(req.params.id)
+            if(isNaN(id)) throw new HttpException(400,'Invalid offer ID')
+
             const deletedOffer = await OfferService.delete(id)
-            res.status(201).json({message: 'Offer deleted successfully'})   
+            res.status(201).json(deletedOffer)   
         }catch(error){
             next(error)
         }
@@ -50,7 +55,10 @@ export class OfferController{
         try{
             const offerData = req.body
             const id = Number.parseInt(req.params.id)
+            if(isNaN(id)) throw new HttpException(400,'Invalid Offer ID')
+
             const newOffer = await OfferService.update(id,offerData)
+            res.status(200).json(newOffer)
         }catch(error){
             next(error)
         }
@@ -58,9 +66,13 @@ export class OfferController{
 
     static async rate(req:Request, res:Response, next:NextFunction){
         try{
-            const {value} = req.body
             const id = Number.parseInt(req.params.id)
+            if(isNaN(id)) throw new HttpException(400,'Invalid offer ID')
+                
+            const {value} = req.body
             const userId = req.user.id
+            if(!userId) throw new HttpException(400, 'User creator ID is required')
+
             await OfferService.rate(userId, id, value)
             res.status(200).json({message: 'Offer rate succesfully'})
         }catch(error){
